# Speed up jump generation in trade routes