Spend a bit more effort wringing out route/jump generation in **TradeCalculation**.

Under profiling, after these changes, raw route generation over imperial sectors takes up over 85% of overall jump-gen time and **base_route_ranges()** takes up around 10%.

Over full-orchestra, jump generation at status quo takes 18 min 42 s (1122 s), while as at "Trim redundant max_dist check...", jump generation over full-orchestra takes 17 min 24 s (1044 s) - a 6.9% reduction.

I'm not sure what other gains are achievable in jump generation, given raw route generation and **base_route_ranges()** are both fairly well tuned for speed.